+++ src/compiler/cpp_generator.cc
@@ -1655,7 +1655,7 @@ std::string GetSourceIncludes(grpc_generator::File* file,
         "grpcpp/support/client_callback.h",
         "grpcpp/support/message_allocator.h",
         "grpcpp/support/method_handler.h",
-        "grpcpp/impl/codegen/rpc_service_method.h",
+        "grpcpp/impl/rpc_service_method.h",
         "grpcpp/impl/codegen/server_callback.h",
         "grpcpp/impl/codegen/server_callback_handlers.h",
         "grpcpp/impl/codegen/server_context.h",
